requirements.txt
environs==11.0.0
python-telegram-bot==13.*
google-cloud-dialogflow==2.34.0
vk_api==11.9.9